Refactor some imports
This will make it easier to avoid a circular dependency
in a commit I'm working on. Also just good practice to
try to import from the most specific module possible.
