+++ shiny/experimental/ui/_tooltip.py
@@ -5,10 +5,9 @@ from typing import Literal, Optional
 
 from htmltools import Tag, TagAttrs, TagAttrValue, TagChild, TagList, tags
 
-from ... import Session
 from ..._namespaces import resolve_id_or_none
 from ..._utils import drop_none
-from ...session import require_active_session
+from ...session import Session, require_active_session
 
 # from ._color import get_color_contrast
 from ._utils import consolidate_attrs
